cheri: drivers: Use uintptr_t for device page allocations
As with the normal page allocations use uintptr_t when allocating
pages with devm_get_free_pages(). Most of the callers should have
been fixed by the FixPtrToULongCast clang-tidy check.

+++ drivers/base/devres.c
@@ -1139,7 +1139,7 @@ devm_kmemdup_const(struct device *dev, const void *src, size_t len, gfp_t gfp)
 EXPORT_SYMBOL_GPL(devm_kmemdup_const);
 
 struct pages_devres {
-	unsigned long addr;
+	uintptr_t addr;
 	unsigned int order;
 };
 
@@ -1171,7 +1171,7 @@ static void devm_pages_release(struct device *dev, void *res)
  * Address of allocated memory on success, 0 on failure.
  */
 
-unsigned long devm_get_free_pages(struct device *dev,
+uintptr_t devm_get_free_pages(struct device *dev,
 				  gfp_t gfp_mask, unsigned int order)
 {
 	struct pages_devres *devres;
@@ -1205,7 +1205,7 @@ EXPORT_SYMBOL_GPL(devm_get_free_pages);
  * Free memory allocated with devm_get_free_pages(). Unlike free_pages,
  * there is no need to supply the @order.
  */
-void devm_free_pages(struct device *dev, unsigned long addr)
+void devm_free_pages(struct device *dev, uintptr_t addr)
 {
 	struct pages_devres devres = { .addr = addr };
